// projects/ksk/dynamic-form/src/lib/interface/form-controller.enum.ts
export enum FORM_CONTROLLER_TYPE {
    EMAIL = "email",
    COLOR = "color",
    DATE = "date",
    DATETIME = "datetime",
    MONTH = "month",
    NUMBER = "number",
    PASSWORD = "password",
    SEARCH = "search",
    TEL = "tel",
    TEXT = "text",
    TIME = "time",
    URL = "url",
    WEEK = "week",
    SELECT = "select",
    RADIO = "radio",
    CHIPS = "chips",
}